feat: add animated thumbnails to thumbnail.go
A lot of duplicate code between the functions, needs a refactor.
