+++ pkg/thumbnail/thumbnail.go
@@ -11,7 +11,7 @@ import (
 
 var (
 	supportedThumbnails = []string{"image/jpeg", "image/png", "image/gif", "image/webp"}
-	animatedThumbnails  = []string{"image/gif", "image/webp"}
+	animatedThumbnails  = []string{"image/gif", "image/webp", "image/avif"}
 
 	cfg config.APIConfig
 )
@@ -36,6 +36,10 @@ func Shutdown() {
 func BuildStaticThumbnail(inputBuf []byte, resp *http.Response) ([]byte, error) {
 	image, err := vips.NewImageFromBuffer(inputBuf)
 
+	if err != nil {
+		return []byte{}, fmt.Errorf("could not load image from url: %s", resp.Request.URL)
+	}
+
 	// govips has the height & width values in int, which means we're converting uint to int.
 	maxThumbnailSize := int(cfg.MaxThumbnailSize)
 
@@ -47,10 +51,6 @@ func BuildStaticThumbnail(inputBuf []byte, resp *http.Response) ([]byte, error) 
 
 	importParams := vips.NewImportParams()
 
-	if err != nil {
-		return []byte{}, fmt.Errorf("could not load image from url: %s", resp.Request.URL)
-	}
-
 	image, err = vips.LoadThumbnailFromBuffer(inputBuf, maxThumbnailSize, maxThumbnailSize, vips.InterestingNone, vips.SizeDown, importParams)
 
 	if err != nil {
@@ -59,6 +59,51 @@ func BuildStaticThumbnail(inputBuf []byte, resp *http.Response) ([]byte, error) 
 	}
 
 	outputBuf, _, err := image.ExportNative()
+
+	if err != nil {
+		return []byte{}, fmt.Errorf("could not export image from url: %s", resp.Request.URL)
+	}
+
+	return outputBuf, nil
+}
+
+func BuildAnimatedThumbnail(inputBuf []byte, resp *http.Response) ([]byte, error) {
+	image, err := vips.NewImageFromBuffer(inputBuf)
+
+	if err != nil {
+		return []byte{}, fmt.Errorf("could not load image from url: %s", resp.Request.URL)
+	}
+
+	// govips has the height & width values in int, which means we're converting uint to int.
+	maxThumbnailSize := int(cfg.MaxThumbnailSize)
+
+	// Only resize if the original image has bigger dimensions than maxThumbnailSize
+	if image.Width() <= maxThumbnailSize && image.Height() <= maxThumbnailSize {
+		// We don't need to resize image nor does it need to be passed through govips.
+		return inputBuf, nil
+	}
+
+	importParams := vips.NewImportParams()
+	format := image.Format()
+
+	// n=-1 is used for animated images to make sure to get all frames and not just the first one.
+	if format == vips.ImageTypeGIF || format == vips.ImageTypeWEBP || format == vips.ImageTypeAVIF {
+		importParams.NumPages.Set(-1)
+	}
+
+	image, err = vips.LoadThumbnailFromBuffer(inputBuf, maxThumbnailSize, maxThumbnailSize, vips.InterestingAll, vips.SizeDown, importParams)
+
+	if err != nil {
+		fmt.Println(err)
+		return []byte{}, fmt.Errorf("could not transform image from url: %s", resp.Request.URL)
+	}
+
+	// exportParams := vips.NewWebpExportParams()
+	// exportParams.Quality = 10
+	// outputBuf, _, err := image.ExportWebp(exportParams)
+
+	outputBuf, _, err := image.ExportNative()
+
 	if err != nil {
 		return []byte{}, fmt.Errorf("could not export image from url: %s", resp.Request.URL)
 	}
